adding flags

+++ ex00/BitcoinExchange.cpp
@@ -41,7 +41,7 @@ int BitcoinExchange::getDateId(std::string str)
 bool isNumber(std::string numStr)
 {
     bool dot = false;
-    for (int i = 0; i < numStr.size(); ++i) {
+    for (size_t i = 0; i < numStr.size(); ++i) {
         if (i == 0 && numStr[i] == '-')
             continue;
         if (!dot && numStr[i] == '.') {
