+++ ex01/RPN.cpp
@@ -3,7 +3,7 @@
 bool isNumber(std::string numStr)
 {
     bool dot = false;
-    for (int i = 0; i < numStr.size(); ++i) {
+    for (size_t i = 0; i < numStr.size(); ++i) {
         if (i == 0 && numStr[i] == '-')
             continue;
         if (!dot && numStr[i] == '.') {
